+++ plugin.php
@@ -11,7 +11,7 @@
  * Plugin Name: Rollback Update Failure
  * Author: Andy Fragen, Ari Stathopolous
  * Description: Feature plugin to test plugin/theme update failures and rollback to previous installed packages.
- * Version: 1.3.4.1
+ * Version: 1.3.4.2
  * Network: true
  * License: MIT
  * Text Domain: rollback-update-failure
@@ -314,7 +314,14 @@ class Rollback_Update_Failure {
 	public function move_dir( $from, $to ) {
 		global $wp_filesystem;
 
-		$result = false;
+		$result           = false;
+		$envs_skip_rename = array( false, 'virtualbox' );
+		$skip_rename      = false;
+
+		// Skip if the runtime environment is in the array.
+		if ( ( in_array( $this->wp_get_runtime_environment(), $envs_skip_rename, true ) ) ) {
+			$skip_rename = true;
+		}
 
 		/*
 		 * Skip the rename() call on VirtualBox environments.
@@ -326,8 +333,7 @@ class Rollback_Update_Failure {
 		 * https://www.virtualbox.org/ticket/17971
 		 */
 
-		if ( 'direct' === $wp_filesystem->method
-			&& ( 'virtualbox' !== $this->wp_get_runtime_environment() ) ) {
+		if ( 'direct' === $wp_filesystem->method && ! $skip_rename ) {
 			$wp_filesystem->rmdir( $to );
 
 			$result = @rename( $from, $to );
@@ -628,31 +634,58 @@ class Rollback_Update_Failure {
 	 *
 	 * @since 6.0.0
 	 *
-	 * @return string The current runtime environment type.
+	 * @return string|false The current runtime environment type. false on failure.
 	 */
 	public function wp_get_runtime_environment() {
 		static $current_runtime_env = '';
+		static $runtime_constant    = '';
+		static $runtime_getenv      = '';
 
 		if ( ! defined( 'WP_RUN_CORE_TESTS' ) && $current_runtime_env ) {
 			return $current_runtime_env;
 		}
 
+		// Array of environments used to bypass using `rename()`.
 		$wp_runtime_environments = array( 'virtualbox' );
 
 		// Fetch the runtime environment from a constant, this overrides the global system variable.
 		if ( defined( 'WP_RUNTIME_ENVIRONMENT' ) ) {
 			$current_runtime_env = WP_RUNTIME_ENVIRONMENT;
-		} elseif ( function_exists( 'getenv' ) ) {
+			$runtime_constant    = $current_runtime_env;
+		}
+
+		// Fetch the runtime environment from global system variable.
+		if ( function_exists( 'getenv' ) ) {
 			// Check if the runtime environment variable has been set, if `getenv` is available on the system.
 			$has_runtime_env = getenv( 'WP_RUNTIME_ENVIRONMENT' );
 			if ( false !== $has_runtime_env ) {
 				$current_runtime_env = $has_runtime_env;
+				$runtime_getenv      = $current_runtime_env;
 			}
 		}
 
-		// Make sure the runtime environment is an allowed one, and not accidentally set to an invalid value.
-		if ( ! in_array( $current_runtime_env, $wp_runtime_environments, true ) ) {
-			$current_runtime_env = '';
+		// If set to something in $wp_runtime_environments use it, default is ''.
+		switch ( true ) {
+			case in_array( $runtime_constant, $wp_runtime_environments, true ):
+				$current_runtime_env = $runtime_constant;
+				break;
+			case in_array( $runtime_getenv, $wp_runtime_environments, true ):
+				$current_runtime_env = $runtime_getenv;
+				break;
+			default:
+				$current_runtime_env = '';
+		}
+
+		if ( ( ! empty( $runtime_constant ) && ! empty( $runtime_getenv ) )
+			&& $runtime_constant !== $runtime_getenv
+		) {
+			_doing_it_wrong(
+				__FUNCTION__,
+				__( 'The WP_RUNTIME_ENVIRONMENT environment variable and constant do not match. Fallbacks will be used, which may negatively impact performance.' ),
+				'6.0.0'
+			);
+
+			return false;
 		}
 
 		return $current_runtime_env;